requirements.txt
Flask~=2.3.3
Werkzeug~=2.3.7
pillow~=11.0.0
requests~=2.32.0
CairoSVG~=2.7.1
reportlab~=4.3.1